# backend/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"categoryParentId": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"type": {
						"type": "string",
						"enum": [
							"news",
							"raovat",
							"menu"
						]
					},
					"styleShow": {
						"type": "string",
						"enum": [
							"news1",
							"news2",
							"news3"
						]
					},
					"createdAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"parentCategory": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"Pick_NewsDto.Exclude_keyofNewsDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"userId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"categoryStyle": {
						"type": "string"
					},
					"parentCategoryId": {
						"type": "string"
					},
					"parentCategoryName": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"publish"
						]
					},
					"slug": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tagIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isHotNews": {
						"type": "boolean"
					},
					"author": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"isPublish": {
						"type": "boolean"
					},
					"customId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_NewsDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_": {
				"$ref": "#/components/schemas/Pick_NewsDto.Exclude_keyofNewsDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"PagedNewsItemDto": {
				"$ref": "#/components/schemas/Omit_NewsDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_"
			},
			"PagedList_PagedNewsItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedNewsItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewStatus": {
				"enum": [
					"draft",
					"publish"
				],
				"type": "string"
			},
...
			"NewsDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"userId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"categoryStyle": {
						"type": "string"
					},
					"parentCategoryId": {
						"type": "string"
					},
					"parentCategoryName": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"publish"
						]
					},
					"slug": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"tagIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isHotNews": {
						"type": "boolean"
					},
					"author": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"isPublish": {
						"type": "boolean"
					},
					"customId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"NewsResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/NewsDto"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewsRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"publish"
						]
					},
					"tagIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isHotNews": {
						"type": "boolean"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"RaoVatDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"extraImages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"rawExtraImagePaths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"createdAt": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"websiteUrl": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"view": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"customId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"WebsiteDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"footerContent": {
						"type": "string"
					},
					"facebookUrl": {
						"type": "string"
					},
					"twitterUrl": {
						"type": "string"
					},
					"googleUrl": {
						"type": "string"
					},
					"linkedinUrl": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"WebsiteResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/WebsiteDto"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": true
			},
			"WebsiteRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"logo": {
						"type": "string"
					},
					"footerContent": {
						"type": "string"
					},
					"facebookUrl": {
						"type": "string"
					},
					"twitterUrl": {
						"type": "string"
					},
					"googleUrl": {
						"type": "string"
					},
					"linkedinUrl": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
			"AdvertisementPosition": {
				"type": "string",
				"enum": [
					"top",
					"left",
					"right"
				]
			},
			"AdvertisementStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
			"AdvertisementDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"position": {
						"$ref": "#/components/schemas/AdvertisementPosition"
					},
					"status": {
						"$ref": "#/components/schemas/AdvertisementStatus"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"imageUrl",
					"url",
					"order",
					"position",
					"status",
					"userId"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"Pick_RaoVatDto.Exclude_keyofRaoVatDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"publishDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"userNameUpdated": {
						"type": "string"
					},
					"metaKeyword": {
						"type": "string"
					},
					"view": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"customId": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"extraImages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"rawExtraImagePaths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"facebook": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"websiteUrl": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_RaoVatDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_": {
				"$ref": "#/components/schemas/Pick_RaoVatDto.Exclude_keyofRaoVatDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"PagedRaoVatItemDto": {
				"$ref": "#/components/schemas/Omit_RaoVatDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_"
			},
			"PagedList_PagedRaoVatItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedRaoVatItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CreateCommentRequest": {
				"properties": {
					"newsId": {
						"type": "string"
					},
					"raoVatId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"anonymousEmail": {
						"type": "string"
					},
					"anonymousName": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"Omit_CommentDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_": {
				"$ref": "#/components/schemas/Pick_CommentDto.Exclude_keyofCommentDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"PagedCommentItemDto": {
				"$ref": "#/components/schemas/Omit_CommentDto.content-or-updateAt-or-updatedBy-or-createdAt-or-createdBy_"
			},
			"PagedList_PagedCommentItemDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/PagedCommentItemDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount",
					"page",
					"size"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ChangePasswordRequest": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"name": "authorization",
				"type": "apiKey",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "nguoivietplus-api",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/admin-console/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "UpsertNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NewsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"MANAGE_POST_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewsRequest"
							}
						}
					}
				}
			}
		},
...
		"/admin-console/website": {
			"get": {
				"operationId": "GetWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "UpsertWebsite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WebsiteResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AdminConsole"
				],
				"security": [
					{
						"": [
							"CONFIG_SEO_EDIT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WebsiteRequest"
							}
						}
					}
				}
			}
		},
...
		"/advertisements": {
			"get": {
				"operationId": "GetAdvertisements",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AdvertisementDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"": [
							"ADS_ADD"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"image"
								]
							}
						}
					}
				}
			}
		},
		"/advertisements/{id}": {
			"get": {
				"operationId": "GetAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAdvertisement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdvertisementDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"": [
							"ADS_EDIT"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAdvertisement",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"": [
							"ADS_DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/news": {
			"get": {
				"operationId": "GetNews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedNewsItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "style",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"news1",
								"news2",
								"news3"
							]
						}
					},
					{
						"in": "query",
						"name": "isHotNews",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "menuSlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rao-vat": {
			"get": {
				"operationId": "GetRaoVats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedRaoVatItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "categorySlug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rao-vat/{slug}": {
			"get": {
				"operationId": "GetRaoVatBySlug",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaoVatDto"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/categories": {
			"get": {
				"operationId": "GetCategoriesByType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"raovat",
								"menu",
								"news"
							]
						}
					}
				]
			}
		},
...
		"/comments": {
			"post": {
				"operationId": "AddComment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PagedList_PagedCommentItemDto_"
								}
							}
						}
					}
				},
				"tags": [
					"website"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"default": 30,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/NewStatus"
						}
					},
					{
						"in": "query",
						"name": "raoVatId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "newsId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"website"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"website"
				],
				"security": [
					{
						"": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	]
}